Adds an aspire based example of azure service bus integration with Akka Streams (#2042)
* Added initial aspire app host

* Adds a simple Service Bus example

Adds a basic example demonstrating the usage of Azure Service Bus with Aspire.

This includes setting up a Service Bus instance using the emulator and adding a queue.

Updates Aspire version to 13.0.0.

* Adds Service Bus topic and subscription

Configures the application to use a Service Bus topic with a subscription.

This setup allows for a publish-subscribe messaging pattern,
enhancing the application's ability to handle asynchronous communication.

* wip - hooked up service bus producer

* Adds a ServiceBus consumer project

Adds a ServiceBus consumer project to the simple ServiceBus example.

This allows for a complete end-to-end example, where the producer sends messages to the ServiceBus and the consumer receives and processes them.

* add readme

---------

Co-authored-by: Richard <richard.harding@iamcloud.com>